client side functions

+++ routes/toDosRoutes.js
@@ -1,12 +1,30 @@
 const express = require('express');
 const todos_router = express.Router();
 const todos_function = require('../db/toDos')
+const { logIn } = require('../db/users')
 
-todos_router.get('/get', async (req, res) => {
+const valid = async (req, res, next) => {
+        try {
+                let [username, password, userId] = req.headers.auth.split(':');
+                req.body = { ...req.body, username, password, userId }
+                const isValid = await logIn(req)
+                if (!isValid) {
+                        res.status(401).send('Invalid username or password')
+                        return
+                }
+                else {
+                        req.user = isValid;
+                        next();
+                }
 
+
+        } catch (err) {
+                console.log(err.message);
+        }
+}
+todos_router.get('/get', valid, async (req, res) => {
         try {
                 const data = await todos_function.getToDos(req)
-
                 if (!data) {
                         res.status(404).send('Not Found')
                 }
@@ -18,7 +36,7 @@ todos_router.get('/get', async (req, res) => {
                 res.status(500).send()
         }
 })
-todos_router.get('/getTrue', async (req, res) => {
+todos_router.get('/getTrue', valid, async (req, res) => {
         try {
                 const data = await todos_function.getTrueToDos(req)
                 if (!data) {
@@ -32,7 +50,7 @@ todos_router.get('/getTrue', async (req, res) => {
                 res.status(500).send()
         }
 })
-todos_router.get('/getFalse', async (req, res) => {
+todos_router.get('/getFalse', valid, async (req, res) => {
         try {
                 const data = await todos_function.getFalseToDos(req)
                 if (!data) {
@@ -48,15 +66,20 @@ todos_router.get('/getFalse', async (req, res) => {
 
 
 })
-todos_router.post('/add', async (req, res) => {
+todos_router.post('/add', valid, async (req, res) => {
         try {
-
-                const data = await todos_function.addTodo(req)
-                if (data.affectedRows < 1) {
-                        res.status(404).send('Not Found')
+                if (req.user.userId === req.params.userId) {
+                        const data = await todos_function.addTodo(req)
+                        if (data.affectedRows < 1) {
+                                res.status(404).send('Not Found')
+                        }
+                        else {
+                                res.send('success')
+                        }
                 }
                 else {
-                        res.send('scsess')
+                        res.status(404).send('Not allowed')
+
                 }
         }
         catch {
@@ -64,15 +87,19 @@ todos_router.post('/add', async (req, res) => {
         }
 
 })
-todos_router.put('/updateTitle/:id', async (req, res) => {
+todos_router.put('/updateTitle/:id', valid, async (req, res) => {
         try {
-                const data = await todos_function.updateTodoTitle(req)
-                if (!data) {
-                        res.status(404).send('Not Found')
+
+                if (req.user.userId === req.params.userId) {
+                        const data = await todos_function.updateTodoTitle(req)
+                        if (!data) {
+                                res.status(404).send('Not Found')
+                        }
+                        else {
+                                res.send(data)
+                        }
                 }
-                else {
-                        res.send(data)
-                }
+                else { res.status(404).send('Not allowed') }
         }
         catch {
                 res.status(500).send()
@@ -82,27 +109,33 @@ todos_router.put('/updateTitle/:id', async (req, res) => {
 })
 todos_router.put('/updateComplet/:id', async (req, res) => {
         try {
-                const data = await todos_function.updateTodoComplet(req)
-                if (!data) {
-                        res.status(404).send('Not Found')
+                if (req.user.userId === req.params.userId) {
+                        const data = await todos_function.updateTodoComplet(req)
+                        if (!data) {
+                                res.status(404).send('Not Found')
+                        }
+                        else {
+                                res.send(data)
+                        }
                 }
-                else {
-                        res.send(data)
-                }
+                else { res.status(404).send('Not allowed') }
         }
         catch {
                 res.status(500).send()
         }
 })
-todos_router.delete('/delete/:id', async (req, res) => {
+todos_router.delete('/delete/:id', valid, async (req, res) => {
         try {
-                const data = await todos_function.deleteTodo(req)
-                if (!data) {
-                        res.status(404).send('Not Found')
+                if (req.user.userId === req.params.userId) {
+                        const data = await todos_function.deleteTodo(req)
+                        if (!data) {
+                                res.status(404).send('Not Found')
+                        }
+                        else {
+                                res.send(data)
+                        }
                 }
-                else {
-                        res.send(data)
-                }
+                else { res.status(404).send('Not allowed') }
         }
         catch {
                 res.status(500).send()
